refactor: use thread runtime event buffer owners
